Get rid of compile time warnings in Xtensive.Orm.Web module
